use css for country flags
Microsoft is not supporting any country flag emojis because they are
afraid of upsetting any sovereignty claims. Eventhough the standard
specification is declaring which country flags are expected, Microsoft
is chosing to do something different. There are so many Windows users
that Microsoft is forcing me to adhere to their descisions.